+++ test/template_builder_test.dart
@@ -12,67 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'dart:convert';
+
 @TestOn('vm')
-import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
 import 'package:test/test.dart';
 
-import 'package:test_html_builder/builder.dart';
 import 'package:test_html_builder/src/builder.dart';
 import 'package:test_html_builder/src/config.dart';
 
 void main() {
-  group('testHtmlBuilder factory', () {
-    test('returns TestHtmlBuilder', () {
-      final options = BuilderOptions({
-        'templates': {
-          'test/bar_template.html': ['test/bar/**_test.dart'],
-          'test/foo_template.html': ['**', 'foo'],
-        },
-      });
-      final builder = testHtmlBuilder(options);
-      expect(builder, isNotNull);
-      expect(
-          builder.getTemplateId(makeAssetId('a|test/bar/bar_test.dart')).path,
-          'test/bar_template.html');
-      expect(builder.getTemplateId(makeAssetId('a|test/foo_test.dart')).path,
-          'test/foo_template.html');
-    });
-
-    test('throws StateError if any template is not in `test/`', () {
-      final options = BuilderOptions({
-        'templates': {
-          'lib/bad_template.html': ['test/**_test.dart'],
-        }
-      });
-      expect(
-          () => testHtmlBuilder(options),
-          throwsA(isA<StateError>().having(
-              (e) => e.message, 'message', contains('lib/bad_template.html'))));
-    });
-
-    test('throws StateError if unsupported option key is provided', () {
-      final options = BuilderOptions({
-        'badkey': 'foo',
-      });
-      expect(() => testHtmlBuilder(options), throwsStateError);
-    });
-
-    test('throws StateError if templates format is invalid', () {
-      final options = BuilderOptions({
-        'templates': ['invalid format'],
-      });
-      expect(() => testHtmlBuilder(options), throwsStateError);
-    });
-  });
-
-  group('TestHtmlBuilder', () {
+  group('TemplateBuilder', () {
     test('does nothing if no templates defined', () async {
-      final config = TestHtmlBuilderConfig();
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       await testBuilder(builder, {
+        'a|test/test_html_builder_config.json': '{}',
         'a|test/foo_test.dart': '',
-        'a|test/template.html': '<html><head>{test}</head></html>',
+        'a|test/template.html': '<html><head>{{testScript}}</head></html>',
       }, outputs: {});
     });
 
@@ -80,10 +36,11 @@ void main() {
       final config = TestHtmlBuilderConfig(templates: {
         'test/template.html': ['test/no_match.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       await testBuilder(builder, {
+        'a|test/test_html_builder_config.json': jsonEncode(config),
         'a|test/foo_test.dart': '',
-        'a|test/template.html': '<html><head>{test}</head></html>',
+        'a|test/template.html': '<html><head>{{testScript}}</head></html>',
       }, outputs: {});
     });
 
@@ -91,16 +48,17 @@ void main() {
       final config = TestHtmlBuilderConfig(templates: {
         'test/template.html': ['test/**_test.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       await testBuilder(builder, {
+        'a|test/test_html_builder_config.json': jsonEncode(config),
         'a|test/bar_test.dart': '',
         'a|test/foo_test.dart': '',
-        'a|test/template.html': '<html><head>{test}</head></html>',
+        'a|test/template.html': '<html><head>{{testScript}}</head></html>',
       }, outputs: {
         'a|test/bar_test.html':
-            '''<html><head><link rel="x-dart-test" href="bar_test.dart"><script src="packages/test/dart.js"></script></head></html>''',
+            '''<html><head><link rel="x-dart-test" href="bar_test.dart"></head></html>''',
         'a|test/foo_test.html':
-            '''<html><head><link rel="x-dart-test" href="foo_test.dart"><script src="packages/test/dart.js"></script></head></html>''',
+            '''<html><head><link rel="x-dart-test" href="foo_test.dart"></head></html>''',
       });
     });
 
@@ -110,17 +68,20 @@ void main() {
         'test/a_template.html': ['test/bar_test.dart'],
         'test/b_template.html': ['test/**_test.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       await testBuilder(builder, {
+        'a|test/test_html_builder_config.json': jsonEncode(config),
         'a|test/bar_test.dart': '',
         'a|test/foo_test.dart': '',
-        'a|test/a_template.html': '<html><head><!-- A -->{test}</head></html>',
-        'a|test/b_template.html': '<html><head><!-- B -->{test}</head></html>',
+        'a|test/a_template.html':
+            '<html><head><!-- A -->{{testScript}}</head></html>',
+        'a|test/b_template.html':
+            '<html><head><!-- B -->{{testScript}}</head></html>',
       }, outputs: {
         'a|test/bar_test.html':
-            '''<html><head><!-- A --><link rel="x-dart-test" href="bar_test.dart"><script src="packages/test/dart.js"></script></head></html>''',
+            '''<html><head><!-- A --><link rel="x-dart-test" href="bar_test.dart"></head></html>''',
         'a|test/foo_test.html':
-            '''<html><head><!-- B --><link rel="x-dart-test" href="foo_test.dart"><script src="packages/test/dart.js"></script></head></html>''',
+            '''<html><head><!-- B --><link rel="x-dart-test" href="foo_test.dart"></head></html>''',
       });
     });
 
@@ -128,13 +89,14 @@ void main() {
       final config = TestHtmlBuilderConfig(templates: {
         'test/template.html': ['test/**_test.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       await testBuilder(builder, {
+        'a|test/test_html_builder_config.json': jsonEncode(config),
         'a|test/bar_test.dart': '',
         'a|test/bar_test.custom.html': 'CUSTOM BAR TEST',
         'a|test/foo_test.dart': '',
         'a|test/foo_test.custom.html': 'CUSTOM FOO TEST',
-        'a|test/template.html': '<html><head>{test}</head></html>',
+        'a|test/template.html': '<html><head>{{testScript}}</head></html>',
       }, outputs: {
         'a|test/bar_test.html': 'CUSTOM BAR TEST',
         'a|test/foo_test.html': 'CUSTOM FOO TEST',
@@ -145,8 +107,9 @@ void main() {
       final config = TestHtmlBuilderConfig(templates: {
         'test/template.html': ['test/**_test.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       final logs = recordLogs(() => testBuilder(builder, {
+            'a|test/test_html_builder_config.json': jsonEncode(config),
             'a|test/foo_test.dart': '',
           }));
       expect(
@@ -161,15 +124,16 @@ void main() {
       final config = TestHtmlBuilderConfig(templates: {
         'test/template.html': ['test/**_test.dart'],
       });
-      final builder = TestHtmlBuilder(config);
+      final builder = TemplateBuilder();
       final logs = recordLogs(() => testBuilder(builder, {
+            'a|test/test_html_builder_config.json': jsonEncode(config),
             'a|test/foo_test.dart': '',
             'a|test/template.html': 'MISSING TOKEN',
           }));
       expect(
           logs,
           emits(severeLogOf(allOf(
-            contains('template must contain a `{test}`'),
+            contains('template must contain exactly one `{{testScript}}`'),
             contains('test/template.html'),
           ))));
     });